Some files
I have added the files I have written so far, mainly Product and DatabaseControl. There is also the BilBook class which is essentially empty. I have also added the necessary images we're going to use in the project. If you need a method in a file just write it and I'll implement it.